refactored tests for login feature
