+++ tests/Routes/login.route.test.ts
@@ -1,29 +1,18 @@
-import { connectToDB } from '../../src/Configs/db.config';
-import { UsersModel } from '../../src/Models/users.model';
-import mongoose from 'mongoose';
-import bcrypt from 'bcrypt';
+import { loginBeforeAll } from '../utils/login/login.beforeAll';
+import { loginAfterAll } from '../utils/login/login.afterAll';
+import { app } from '../../src/index';
 import request from 'supertest';
-import { app, server } from '../../src/index';
+import bcrypt from 'bcrypt';
 
 describe('Test of login route', () => {
     let hashedPassword: string;
 
     beforeAll(async () => {
-        await connectToDB();
-
         hashedPassword = await bcrypt.hash('password', 10);
-
-        await UsersModel.create({
-            login: 'login_routes',
-            password: hashedPassword,
-        });
+        await loginBeforeAll('route', hashedPassword);
     });
 
-    afterAll(async () => {
-        await UsersModel.deleteOne({ login: 'login_routes' });
-        await mongoose.disconnect();
-        server.close();
-    });
+    afterAll(loginAfterAll('route'));
 
     it.only('Correct login data', async () => {
         const loginData = {
