+++ tests/Services/login.service.test.ts
@@ -1,27 +1,17 @@
-import { UsersModel } from '../../src/Models/users.model';
 import { checkLoginData } from '../../src/Services/login.service';
-import { connectToDB } from '../../src/Configs/db.config';
+import { loginBeforeAll } from '../utils/login/login.beforeAll';
+import { loginAfterAll } from '../utils/login/login.afterAll';
 import bcrypt from 'bcrypt';
-import mongoose from 'mongoose';
 
 describe('Test of login service', () => {
     let hashedPassword: string;
 
     beforeAll(async () => {
-        await connectToDB();
-
         hashedPassword = await bcrypt.hash('password', 10);
-
-        await UsersModel.create({
-            login: 'login_service',
-            password: hashedPassword,
-        });
+        await loginBeforeAll('service', hashedPassword);
     });
 
-    afterAll(async () => {
-        await UsersModel.deleteOne({ login: 'login_service' });
-        await mongoose.disconnect();
-    });
+    afterAll(loginAfterAll('service'));
 
     it.only('Correct login data', async () => {
         const foundUser = await checkLoginData({
